Allow overriding deployment and container fields
This adds deployment override fields to the VPA operator CRD such that
the user can now specifically override the container resources and
arguments for the containers in the operand deployments managed by the
VPA operator.

We aren't allowing overrides of all the deployment fields for now, just
the ones we currently support, but we're leaving room with the thought
that someday the overrides could expand to cover a nontrivial portion of
the deployment spec.