Use modern controller-gen method for generation
This project was generated with a very old v0.10.0? version of
operator-sdk, and its generators and makefile didn't work properly in
a modern go 1.20 environment.

This replaces the "generate" target and adds a "manifests" target to get
at least the code/CRD generation updated to match a modern operator-sdk
project, and we'll need to come back to migrate the rest of the project
layout and scripts.

With this, "make generate" and "make manifests" will now do what you'd
expect again.