Pull deprecated CVO functions into VPA lib
We were using a couple of resourcemerge utility functions out of the
CVO, and the CVO stopped using them and got rid of them, so we can't use
them from there anymore.

I did check library-go and there didn't appear to be an equivalent
function we could switch to.

We still need them, and there's not a compelling reason to refactor at
this time, so this just pulls them into our own VPA resourcemerge
library until someday when maybe we get rid of them too.