Migrate to new controller-runtime fields
controller-runtime > v0.15.2 changed how the cache namespacing works and
deprecates the Namespace options field, which used to be used under the
hood to spet the namespace for the operator's watches/caches.

This caused some problems because it wasn't a 1:1 swich to the new cache
options -- the new cache does not automatically assume the namespace
during a Get of a NamespacedName, and we apparently had a few spots in
here where the namespace was missing from the NamespacedName (we had it
available, we just didn't supply it).

This adjusts the manager fields to accommodate the new
controller-runtime upgrade and fills in the namespace in a couple of
spots where we previously negleted to do that. The behavior should now
be exactly the same as it was on the older controller-runtime version.