Remove go mod verify from Dockerfile
So because we run our builds in a disconnected environment without
internet access, the "go mod verify" statement will fail trying to
check the module hashes.

There might be a module cache or something as part of the build
environment that we can we can wire up so we can still verify while in
the disconnected environment, but until we figure it out, we want our
builds to still succeed.

This just removes the "go mod verify" command for now to make builds
work again while we figure out of there's a way we can still make it
work.

Signed-off-by: John Kyros <jkyros@redhat.com>