Revert operator package and csv name changes
This commit reverts the operator package name change from vertical-pod-autoscaler-operator back to vertical-pod-autoscaler. This reason for this change is because package names need to be consistent throughout the lifecycle of an operator through various upgrades. Although this doesn't matter much since ART will apply the correct package name from our bundle/manifests/{VPA}.package.yaml file, this will make it easier to run our upstream upgrade tests. This also reverts the CSV name changes because the Operator-SDK tool does not like it when your package name does not match your csv name and this triggers a lot of unexpected errors, and there doesn't seem to be much merit to applying a bunch of hacks to fix them.