Re-enable VPA alm-example in CSV
This commit re-enables the VPA example in the CSV for the OCP console to use. Turns out the VPA is a "supported resource" and operator-sdk adds your supported resource examples to the bundle regardless if it's actually an owned resource or not. This commit simply uncomments the sample, and manually removes it from the bundle folder after generating it with Makefile.
This commit also fixes a bug when deploying the bundle with Make and specifying an OPERAND_IMG.