Updates for 4.20
This commit updates references in the repo from 4.19 to 4.20, except for actually bumping go dependencies.
We need this to satisfy ART builds, but we cannot do a full update for our OpenShift 4.20 dependencices yet
due to the k8s 1.33 rebase not being finished, so we will hold off on golang dep bumps, go1.24 bumps, and
image bumps until then.

Signed-off-by: Max Cao <macao@redhat.com>