Fix version injection regression
When we updated operator-sdk to 1.35 we simplified a lot of the
makefile/dockerfile process, and the version injection accidentally got
removed.

This just re-adds it and make the dockerfile call the build from the
Makefile so we don't have to duplicate the version logic in multiple
places.

This also adds a test to the e2e script so we will fail e2e if this ever
happens again.

Signed-off-by: John Kyros <jkyros@redhat.com>